requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bench==2.8
billiard==4.2.0
boto3==1.29.7
botocore==1.32.7
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.57.0
django-background-tasks==1.2.5
django-celery-beat==2.5.0
django-compat==1.0.15
django-cors-headers==4.3.0
django-multiselectfield==0.1.12
django-restframework==0.0.1
django-storages==1.14.2
django-timezone-field==6.1.0
django-upgrade==1.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.109.2
frappe-bench==5.17.2
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
honcho==1.1.0
httptools==0.6.1
idna==3.4
install==1.3.5
Jinja2==3.0.3
jmespath==1.0.1
kombu==5.3.4
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.8.0.74
Pillow==10.1.0
pipreqs==0.4.13
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.8.0
schedule==1.2.1
semantic-version==2.8.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
tokenize-rt==5.2.0
tomli==2.0.1
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.27.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
yarg==0.1.9
yarl==1.9.2
